feat: python3.11, territory getters for population, living area and houses

+++ population_setler/models/house.py
@@ -5,4 +5,4 @@ class House:
     """todo descr"""
     id : int | None = None
     population: int | None = None
-    living_area: int | None = None
+    living_area: float | None = None
